+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fixed the consistency of the filenames across downloads.
 * Reduced the length of the zip filenames.
 * Bug fix for a bug which appears on staging within `application_generate_zip`.
+* Bug fix to only generate a zip for all job applications if any of the submissions have valid attachments.
 
 
 ### 0.1.0